requirements.txt

uvicorn

soundfile
torchaudio==2.7.0
torch
python-Levenshtein
pydu
pydub
silero-vad
seaborn

psycopg2-binary
spacy

ipython
datasets
python-dotenv

boto3

demucs